requirements.txt
aiohttp
pycryptodome

